refactoring download from UHSLC
