+++ oceanicospy/retrievals/__init__.py
@@ -1,4 +1,3 @@
-from .download_CMDS_data import *
-from .download_ERA5_data import *
 from .download_UHSLC_data import *
-from .download_IDEAM_data import *
+# from .download_CMDS_data import *
+# from .download_ERA5_data import *
